requirements/requirements.txt
colossalai==0.4.0  # TODO: build from source with `git checkout v0.4.0 && BUILD_EXT=1 pip install .`
mmengine==0.10.5
pandas>=2.0.3
timm==0.9.16
rotary_embedding_torch==0.5.3
ftfy>=6.2.0 # for t5
diffusers==0.27.2 # for vae
accelerate==0.29.2 # for t5
av>=12.0.0,<14.0.0 # for video loading
numpy<2.0.0
huggingface_hub==0.25.0

# [gradio]
gradio>=4.26.0
spaces>=0.28.3

# [notebook]
ipykernel>=6.29.4
ipywidgets>=8.1.2

# [training]
wandb>=0.17.0
tensorboard>=2.14.0
pandarallel>=1.6.5
pyarrow>=16.1.0 # for parquet
librosa==0.9.2 # for audio
soundfile==0.12.1
audioflux==0.1.9
scipy<=1.14.1
iopath==0.1.10
pytorchvideo==0.1.5
transformers==4.39.3
peft==0.12.0

# [dev]
pre-commit>=3.5.0
openai
